Refactor SlimWizard to improve command handling and modularity
Introduced `CommandAdapter` to streamline enhanced command logic and refactored wizard steps to simplify creation and navigation. Enhanced reactivity in `Wizard` class and addressed command initialization issues for better maintainability.
